refactor(orders): consolidate address input types and update order management
- Replace ShippingAddressInput with AddressInput to match GraphQL schema
- Update PlaceOrderInput to use unified AddressInput type
- Modify OrderResolver to handle new address field structure
- Update PaymentIntentResponse type consistency
- Enhance CalendarOrder model for better order management
- Update OrderResolverTest to reflect schema changes
- Mark order management task as completed in CodeMachine

+++ src/main/java/villagecompute/calendar/api/graphql/types/PaymentIntentResponse.java
@@ -24,7 +24,7 @@ public class PaymentIntentResponse {
     public String clientSecret;
 
     @Description("Amount in cents (USD) - multiply dollar amount by 100")
-    public Long amount;
+    public Integer amount;
 
     @Description("Associated calendar ID (before order is created)")
     public UUID calendarId;
@@ -48,7 +48,7 @@ public class PaymentIntentResponse {
     public static PaymentIntentResponse fromCheckoutSession(
         String sessionId,
         String checkoutUrl,
-        Long amountInCents,
+        Integer amountInCents,
         UUID calendarId,
         Integer quantity
     ) {
